Add generic condition to npc location data & resolution
