+++ Assets/Editor/CustomInspector/TemporaryConditionalObjectEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5b0d22b2a53e0b9469dfb4ae39238de2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
