+++ Assets/Scripts/Data/DataModels/NPCs/NPCLocation.cs
@@ -10,6 +10,9 @@ public class NPCLocation
     [field: SerializeField]
     public Vector3 EveningStationPosition { get; private set; }
 
+    [field: SerializeField]
+    public Condition[] Conditions { get; private set; }
+
     public Vector3 GetPositionByPhase(TimeOfDay.Phase phase)
     {
         return phase == TimeOfDay.Phase.Morning ? MorningStationPosition : EveningStationPosition;
